Split up beer_controller.js into separate routes. Mapped to those new routes on server.js.
